update frame dependency completion monitoring

+++ foodx_devops_tools/deploy_me/_state.py
@@ -14,8 +14,8 @@ class PipelineCliOptions:
     """Pipeline configuration options provided from CLI."""
 
     enable_validation: bool
-    monitor_sleep_seconds: int
-    wait_timeout_seconds: int
+    monitor_sleep_seconds: float
+    wait_timeout_seconds: float
 
 
 @enum.unique
